Update CRUD Full

+++ resources/js/Pages/Store.vue
@@ -4,7 +4,7 @@
     <div class="card">
       <div class="card-header pb-0">
         <div class="d-flex justify-content-between">
-          <h4 class="card-title mg-b-0">STRIPED ROWS</h4>
+          <h4 class="card-title mg-b-0">ລາຍການ ສະຕ໋ອກສິນຄ້າ</h4>
           <span>
        
             
@@ -39,12 +39,11 @@
             </button>
           </span>
         </div>
-        <p class="tx-12 tx-gray-500 mb-2">
-          Example of Valex Striped Rows.. <a href="">Learn more</a>
-        </p>
+      
       </div>
       <div class="card-body">
-    
+        
+        
       
         <div class="form-store row" v-if="FormShow">
           <div class="col-md-3">ອັບໂຫຼດຮູບ</div>
@@ -95,32 +94,49 @@
           </div>
           
         </div>
-      </div>
+      
 
       <div class="table-responsive" v-if="!FormShow">
-        <table class="table table-striped mg-b-0 text-md-nowrap">
+        <div class="row">
+          <div class="col-md-6"></div>
+          <div class="col-md-6">
+            <input
+                type="text"
+                class="form-control"
+                placeholder="ຄົ້ນຫາສິນຄ້າ..."
+                v-model="search"
+                @keyup.enter="get_data()"
+              />
+          </div>
+        </div>
+        
+        <table class="table table-striped mg-b-0 text-md-nowrap mt-4 border">
           <thead>
             <tr>
               <th>ID</th>
               <th>ຊື່ສິນຄ້າ</th>
-              <th>ຈຳນວນ</th>
-              <th>ລາຄາຊື້</th>
-              <th>ລາຄາຂາຍ</th>
-              <th>ຈັດການ</th>
+              <th width="80" class=" text-center">ຈຳນວນ</th>
+              <th class=" text-center">ລາຄາຊື້</th>
+              <th class=" text-center">ລາຄາຂາຍ</th>
+              <th width="120" class=" text-center">ຈັດການ</th>
             </tr>
           </thead>
           <tbody>
-            <tr v-for="list in FormData" :key="list.id">
+            <tr v-if="FormData.data.length == 0">
+              <td colspan="6" class=" text-center"> <h2> ບໍ່ມີຂໍ້ມູນ </h2></td>
+            </tr>
+            <tr v-for="list in FormData.data" :key="list.id">
               <th>{{ list.id }}</th>
               <td>{{ list.name }}</td>
-              <td>{{ list.amount }}</td>
-              <td>{{ formatPrice(list.price_buy) }}</td>
-              <td>{{ formatPrice(list.price_sell) }}</td>
-              <td>
+              <td class=" text-center">{{ list.amount }}</td>
+              <td class=" text-right">{{ formatPrice(list.price_buy) }}</td>
+              <td class=" text-right">{{ formatPrice(list.price_sell) }}</td>
+              <td class=" text-center">
                 <div class="btn-icon-list">
                   <button class="btn btn-primary btn-icon" @click="edit_store(list.id)">
                     <i class="fe fe-edit"></i>
-                  </button>
+                  </button> 
+                  <!-- <button class="btn btn-danger btn-icon" @click="showAlert()"> -->
                   <button class="btn btn-danger btn-icon" @click="delete_store(list.id)">
                     <i class="fe fe-trash-2"></i>
                   </button>
@@ -129,6 +145,8 @@
             </tr>
           </tbody>
         </table>
+        <pagination :pagination="FormData" :offset="4" @paginate="get_data($event)"/>
+      </div>
       </div>
       <!-- bd -->
     </div>
@@ -138,11 +156,14 @@
 </template>
 
 <script>
+import Pagination from '../components/Pagination.vue';
 export default {
+  components: { Pagination },
   name: "MiniPosStore",
 
   data() {
     return {
+      search:'',
       options: {
               ///prefix: '₭ ',
               numeral: true,
@@ -168,13 +189,43 @@ export default {
   },
 
   methods: {
+    showAlert() {
+      // Use sweetalert2
+      //this.$swal('Hello Vue world!!!');
+
+      this.$swal.fire({
+          title: 'Are you sure?',
+          text: "You won't be able to revert this!",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+          if (result.isConfirmed) {
+            this.$swal.fire(
+              'Deleted!',
+              'Your file has been deleted.',
+              'success'
+            )
+          }
+        });
+    },
     formatPrice(value) {
 			let val = (value / 1).toFixed(0).replace(",", ".");
 			return val.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ",");
 		},
-    get_data(){
-      this.$axios.get('/api/store').then((response)=>{
+    get_data(page){
+  
+      /*this.$axios.get('/api/store').then((response)=>{
             console.log(response.data);
+            this.FormData = response.data;
+            }).catch((error)=>{
+                console.log(error)
+            });*/
+      
+      this.$axios.get(`/api/store?page=${page}&s=${this.search}`).then((response)=>{
+            //console.log(response.data);
             this.FormData = response.data;
             }).catch((error)=>{
                 console.log(error)
@@ -219,17 +270,46 @@ export default {
         //let index = this.FormData.map((i)=>i.id).indexOf(id);
         //this.FormData.splice(index,1);
 
-        this.$axios.delete(`/api/store/delete/${id}`).then((response)=>{
+        this.$swal.fire({
+          title: 'ທ່ານແນ່ໃຈບໍ່?',
+          text: "ທີ່ຈະລຶບຂໍ້ມູນນີ້!",
+          icon: 'question',
+          showCancelButton: true,
+          confirmButtonColor: '#3085d6',
+          cancelButtonColor: '#d33',
+          confirmButtonText: 'ຕົກລົງ',
+          cancelButtonText:'ຍົກເລີກ'
+        }).then((result) => {
+          if (result.isConfirmed) {
+
+            this.$axios.delete(`/api/store/delete/${id}`).then((response)=>{
               if(response.data.success){
                     this.get_data();
-                    alert(response.data.message);
+                    //alert(response.data.message);
+                    this.$swal.fire(
+                    'ສຳເລັດ!',
+                    response.data.message,
+                    'success'
+                  );
               } else {
-                  alert('ລຶບຂໍ້ມູນ ບໍ່ສຳເລັດ:' + response.data.message);
+                  //alert('ລຶບຂໍ້ມູນ ບໍ່ສຳເລັດ:' + response.data.message);
+
+                  this.$swal.fire(
+                    'ຜິດຜາດ!',
+                    response.data.message,
+                    'error'
+                  );
               }
               
             }).catch((error)=>{
                 console.log(error)
             });
+
+            
+          }
+        });
+
+        
     },
     save_data() {
 
@@ -301,6 +381,13 @@ export default {
   },
   created(){
     this.get_data();
+  },
+  watch:{
+    "search"(){
+        if(this.search == ""){
+          this.get_data();
+        }
+    }
   }
 };
 </script>
